moved internal headers to internal folder, for clearer doc
