+++ doc/01.overview.md
@@ -1,7 +1,7 @@
 # Overview
 
 o80 (pronounced 'oh eighty') is a tool for synchronizing processes while organizing exchange of information between them.
-The information exchanged are commands for computing (robotic) desired states and observations, where state and observation are user class (and o80 classes are templated over them).
+The information exchanged are commands for computing (robotic) desired states and observations, where state and observation are (almost) arbitrary user declared and defined classes (o80 classes are templated over them).
 
 Here is a basic usage of o80, using python:
