+++ internal/o80_internal/command.hxx
