+++ internal/o80_internal/command_status.hpp
