+++ internal/o80_internal/command_type.hpp
