+++ internal/o80_internal/time_stamp.hpp
